cleanup and documents
