+++ letterboxd/me_test.go
@@ -1,15 +1,15 @@
 package letterboxd
 
-import (
-	"testing"
-)
+import "testing"
 
 // TODO: This isn't testing correctly. Needs a mocked endpoint
-func TestGetAuthToken(t *testing.T) {
+func TestGetMe(t *testing.T) {
 	c := NewClientFromKey("FAKE_KEY", "FAKE_SECRET")
-	at, err := c.GetAuthToken("USERNAME", "PASSWORD")
+
+	at, _ := c.GetAuthToken("USERNAME", "PASSWORD")
+	r, err := c.GetMe(at.AccessToken)
 
 	if err != nil {
-		t.Error(at)
+		t.Error(r)
 	}
 }
