+++ letterboxd/request_test.go
@@ -6,11 +6,9 @@ import (
 )
 
 func TestBuild(t *testing.T) {
-	// // precomputed request signature. haven't figured out how to test this func
-	// eSig := ""
-	// eURL := "https://api.letterboxd.com/api/v0/fake?apikey=FAKE_KEY&nonce=&timestamp=&signature="
 	m := "GET"
 	e := "/fake"
+	d := ""
 	b := ""
 
 	c := NewClientFromKey("FAKE_KEY", "FAKE_SECRET")
@@ -18,6 +16,7 @@ func TestBuild(t *testing.T) {
 	r := builder.
 		Method(m).
 		Endpoint(e).
+		Data(d).
 		Body(b).
 		Build(c)
 
@@ -32,25 +31,10 @@ func TestBuild(t *testing.T) {
 	if r.body != b {
 		t.Errorf(errorMsg("Incorrect body generated.", b, r.body))
 	}
-
-	// if r.signature != eSig {
-	// 	t.Errorf(errorMsg("Incorrect signature generated.", eSig, r.signature))
-	// }
-
-	// if r.url != eURL {
-	// 	t.Errorf(errorMsg("Incorrect URL generated.", eURL, r.url))
-	// }
 }
 
+// TODO: This definitely needs to be tested specifically. Not sure how to get a baseline "correct" signature because of the timestamp and UUID requirements
 func TestBuildSignature(t *testing.T) {
-	// // precomputed request signature. haven't figured out how to test this func
-	// eSig := ""
-	// c := NewClientFromKey("FAKE_KEY", "FAKE_SECRET")
-	// r := c.NewRequest("GET", "/fake", "")
-
-	// if r.signature != eSig {
-	// 	t.Errorf(errorMsg("Incorrect signature generated.", eSig, r.signature))
-	// }
 }
 
 func errorMsg(m string, w string, g string) string {
